[Refactor] ViewController, MainView 분리
